--- src/main/java/ru/sber/projectATM/core/helperСlass/Handbook.java
@@ -1,17 +1,0 @@
-package ru.sber.projectATM.core.helperСlass;
-
-public class Handbook {
-
-    public enum Currency {
-        RUR,
-        USD,
-        EUR
-    }
-
-    public enum Status {
-        SUCCESS,
-        FAILED,
-        WAITED
-    }
-
-}
