+++ src/main/java/ru/sber/projectATM/core/helperСlass/accounting/Balance.java
@@ -1,0 +1,11 @@
+package ru.sber.projectATM.core.helperСlass.accounting;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class Balance {
+    private Double amount;
+    private Currency currency;
+}
