+++ src/main/java/ru/sber/projectATM/core/helperСlass/accounting/Currency.java
@@ -1,0 +1,7 @@
+package ru.sber.projectATM.core.helperСlass.accounting;
+
+public enum Currency {
+    RUR,
+    USD,
+    EUR
+}
