+++ src/main/java/ru/sber/projectATM/core/helperСlass/accounting/Status.java
@@ -1,0 +1,7 @@
+package ru.sber.projectATM.core.helperСlass.accounting;
+
+public enum Status {
+    SUCCESS,
+    FAILED,
+    WAITED
+}
